hashing: move to repository package
