+++ internal/repository/hashing/writer.go
