+++ internal/repository/packer_manager.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/restic/restic/internal/backend"
 	"github.com/restic/restic/internal/errors"
-	"github.com/restic/restic/internal/hashing"
+	"github.com/restic/restic/internal/repository/hashing"
 	"github.com/restic/restic/internal/restic"
 
 	"github.com/restic/restic/internal/crypto"
